Updated Login Api and updated GetAllProperties Api
